<LeetCode202109> 98. Validate Binary Search Tree

+++ 0098_Validate_Binary_Search_Tree/Solution2.cpp
@@ -2,13 +2,13 @@
 #include <iostream>
 #include <climits>
 
-bool check(TreeNode* root, int minVal, int maxVal) {
+bool check(TreeNode* root, long minVal, long maxVal) {
 	if (!root) return true;
 
 	if (root->val <= minVal || root->val >= maxVal) {
 		return false;
 	}
-	return check(root->left, INT_MIN, root->val) && check(root->right, root->val, INT_MAX);
+	return check(root->left, LONG_MIN, root->val) && check(root->right, root->val, LONG_MAX);
 }
 
 bool Solution::isValidBST2(TreeNode* root) {
